requirements.txt
openai
slack-bolt
python-dotenv
